requirements.txt
glfw
PyOpenGL
numpy
pyrr
Pillow
scipy